==> code/03b_homepage_data_integration.ipynb <==
# Homepage data integration - **final stage**

As part of the ** final stage ** the tools' homepage data is fetched and saved.

## Imports

import pandas as pd
import requests

## Load tools from consolidated stage

Try to get homepage data for all the tools present in consolidated tool table.

Only tools can be considered where a homepage URL is available.

df = pd.read_csv("data/04_consolidated/tools.csv")

df.shape

df.loc[df["homepage_url"].notna()].shape

## Fetch homepage data HTTP GET requests

A simple approach using simple HTTP GET requests and saving the respective response text.

Response status, response text for each tool are combined in a DataFrame.

homepage_responses=[]
for intex, row in df.loc[df["homepage_url"].notna()].iterrows():
    try:
        response = requests.get(row["homepage_url"])
        status=response.status_code
        responste_text = response.text
        homepage_responses.append({"tool":row["id"],"status_code":response.status_code,"text":response.text})
    except Exception as e:
        print(row["id"])
        homepage_responses.append({"tool":row["id"],"error":e})


df=pd.DataFrame(homepage_responses).rename(columns={"tool":"id"})
df.shape

df_no_exception=df[df["error"].isna()]
df_no_exception.shape

df_status_code_200=df_no_exception[df_no_exception["status_code"]==200]
df_status_code_200.shape

### Save result of homepage data integration

Save tools and homepage data where fetching was successful (HTTP 200 response).

Result is saved in **final** data folder.

df_status_code_200.to_csv("data/05_final/homepages.csv",index=False)
